SECURITY FIX for setfenv and getfenv
By themselves, getfenv and setfenv are safe. When scripts use getfenv to grab an unsecure environment it is wrapped and secure. The issue is that when that environment is passed back into setfenv then the environment is unwrapped and applied, potentially overwriting a secure environment. This update makes setfenv check if it is overwriting a secure environment and if so then it will write a secure environment. In addition, getfenv and setfenv will now give the expected behavior of 1 getting / setting the caller's environment instead of the sandboxes environment.